Implement Phase 2 of UNHOX microkernel with user thread support and IPC enhancements
- Added `thread_create_user` function to create user threads with proper stack and entry point.
- Implemented `user_entry_trampoline` in assembly for transitioning to user mode.
- Updated `paging_create_task_pml4` to allocate private page tables for user tasks, ensuring proper user access rights.
- Refactored bootstrap server to operate as a dedicated kernel thread, handling IPC messages for service registration and lookup.
- Introduced new IPC message structures for registration and lookup, enhancing communication between services.
- Updated registry management to support new IPC-based service registration and lookup mechanisms.
- Enhanced the build script to include user space initialization and ensure proper kernel image paths.
- Added detailed kernel architecture and heritage documentation to reflect the design philosophy and lineage of UNHOX.
- Conducted milestone tests for IPC functionality, confirming successful message passing and user task execution.

+++ servers/bootstrap/registry.c
@@ -1,14 +1,13 @@
 /*
  * servers/bootstrap/registry.c — Service name → port registry for UNHOX
  *
- * The registry is a simple array-based map from service names to port
- * handles.  In a real system this would be backed by IPC (the bootstrap
- * server would receive registration and lookup messages on its bootstrap
- * port).  For Phase 1 we implement it as a direct C API callable from
- * the kernel for testing.
+ * The registry maps service names (strings) to port handles (ipc_port *,
+ * carried as uint64_t to avoid a kernel header dependency in the .h).
  *
- * TODO (Phase 2): Convert to a message-handling loop that receives
- *                 MIG-generated bootstrap RPCs on the bootstrap port.
+ * In a real system the bootstrap server would receive registration and
+ * lookup requests as Mach messages on its bootstrap port and update this
+ * registry in response.  For UNHOX Phase 2, the message dispatch happens
+ * in bootstrap.c; this file is the backing store.
  *
  * Reference: OSF MK bootstrap/bootstrap_defs.h for the message format.
  */
@@ -42,7 +41,7 @@ static void bs_strncpy(char *dest, const char *src, int n)
 
 struct service_entry {
     char        name[BOOTSTRAP_NAME_MAX];
-    uint32_t    port;
+    uint64_t    port;       /* ipc_port * carried as opaque integer */
     int         active;     /* 1 = registered and active */
     int         checked_in; /* 1 = server has claimed this slot */
 };
@@ -65,7 +64,7 @@ void bootstrap_init(void)
     registry_count = 0;
 }
 
-int bootstrap_register(const char *name, uint32_t port)
+int bootstrap_register(const char *name, uint64_t port)
 {
     if (!name)
         return BOOTSTRAP_NOT_PRIVILEGED;
@@ -94,7 +93,7 @@ int bootstrap_register(const char *name, uint32_t port)
     return BOOTSTRAP_NO_MEMORY;
 }
 
-int bootstrap_lookup(const char *name, uint32_t *out_port)
+int bootstrap_lookup(const char *name, uint64_t *out_port)
 {
     if (!name || !out_port)
         return BOOTSTRAP_UNKNOWN_SERVICE;
@@ -109,7 +108,7 @@ int bootstrap_lookup(const char *name, uint32_t *out_port)
     return BOOTSTRAP_UNKNOWN_SERVICE;
 }
 
-int bootstrap_checkin(const char *name, uint32_t *out_port)
+int bootstrap_checkin(const char *name, uint64_t *out_port)
 {
     if (!name || !out_port)
         return BOOTSTRAP_UNKNOWN_SERVICE;
